[Refactor]: Update `e3sm_diags_driver.run_diags` as `CoreParameter` method (#742)
- Update VSCode workspace settings with extension and debugger configs
